dia 09: novo exemplo scanner

+++ src/main/java/com/linuxtips/descomplicandojavaespring/day08/PedidoScanner.java
@@ -17,11 +17,14 @@ public class PedidoScanner {
 
         System.out.println("Digite o valor do produto");
 
-        Double valor = scanner.nextDouble();
+        Integer valor = scanner.nextInt();
 
-        double total = valor * quantidade;
+        int total = valor * quantidade;
 
         System.out.println("O valor total da sua compra é  " + total);
+
+
+
 
 
     }
